Crate roboime_next

A library for abstracting the most common interfaces and filtering for making an AI for the RoboCup's SSL robot soccer game.

The main purpose is to provide an stdio based interface for building an AI. The usage of this library is useful for building or customizing a user interface around that concept.

Examples

The following will make a CLI program that will run an executable called demo-ai for it to play on a running grSim (or any simulator protocol compatible).

Note that there is no referee in this example.

use std::process::Command;
use roboime_next::prelude::*;
use roboime_next::{ai, grsim};

let mut grsim = grsim::Builder::new()
    .vision_addr("224.5.23.2:11002").unwrap()
    .grsim_addr("127.0.0.1:20011").unwrap()
    .build().unwrap();

let mut ai = ai::Builder::new(Command::new("./demo-ai"))
    .color(Yellow)
    .build().unwrap();

grsim.wait_for_geom().unwrap();
let mut ai = ai.init(&grsim).unwrap();

loop {
    grsim.recv_state().unwrap();
    let cmd = ai.update(&grsim).unwrap();
    grsim.send_command(cmd).unwrap();
}

To see that in action you will need a demo-ai, and a running grSim with matching settings (which could very well be running on another machine, just change the grsim_addr).

Modules

ai

base

game
This module is mostly concerned with the game state.

grsim

prelude

sim
A built-in simulator, with referee support.

Structs

Error
The error type for this crate.

Enums

ErrorKind
General category of errors.

Type Definitions

Result
A specialized Result type for this crate.