Trait roboime_next::game::Geometry

pub trait Geometry {
    fn field_length(&self) -> f32;
    fn field_width(&self) -> f32;
    fn goal_width(&self) -> f32;
    fn center_circle_radius(&self) -> f32;
    fn defense_radius(&self) -> f32;
    fn defense_stretch(&self) -> f32;
}

Specifications of the field geometry

This specifications intend to be minimal, some fields that are available from the vision are left out on purpose.

Required Methods

fn field_length(&self) -> f32

fn field_width(&self) -> f32

fn goal_width(&self) -> f32

fn center_circle_radius(&self) -> f32

fn defense_radius(&self) -> f32

fn defense_stretch(&self) -> f32